+++ code/server/SmartHome/src/cn/fuego/smart/home/domain/UserConcentrator.java
@@ -28,7 +28,7 @@ public class UserConcentrator implements PersistenceObject
 	 */
 	private static final long serialVersionUID = 1L;
 	private Integer userID;
-	private Integer concentratorID;
+	private Long concentratorID;
 	private Integer operate;  //0,read 1 modify 2 delete 3 all
 	public Integer getUserID()
 	{
@@ -38,11 +38,11 @@ public class UserConcentrator implements PersistenceObject
 	{
 		this.userID = userID;
 	}
-	public Integer getConcentratorID()
+	public Long getConcentratorID()
 	{
 		return concentratorID;
 	}
-	public void setConcentratorID(Integer concentratorID)
+	public void setConcentratorID(Long concentratorID)
 	{
 		this.concentratorID = concentratorID;
 	}
